+++ src/features/diary/diaryList/components/DiaryCard.tsx
@@ -1,10 +1,10 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import { File } from '@/features/diary/apis/interfaces.ts';
-import * as style from '@/features/diary/components/diaryList/DiaryCard.css';
+import { File } from '@/features/diary/diaryList/apis/interfaces';
+import * as style from '@/features/diary/diaryList/components/DiaryCard.css';
 import earth from '@/static/diary/earth2.svg';
 import calendar from '@/static/diary/calendar.svg';
-import DiaryCardDropdown from '@/features/diary/components/diaryList/DiaryCardDropdown.tsx';
+import DiaryCardDropdown from '@/features/diary/diaryList/components/DiaryCardDropdown.tsx';
 import { MouseEvent } from 'react';
 
 export default function DiaryCard({ file }: { file: File }) {
